Implementing AdaptSize with C++ instead of C for better data structure
